requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
kiwisolver==1.4.8
matplotlib==3.10.0
MouseInfo==0.1.3
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytweening==1.2.0
rubicon-objc==0.5.0
six==1.17.0
